+++ app/Http/Controllers/UsersController.php
@@ -160,6 +160,18 @@ class UsersController extends Controller
      */
     public function delete($id)
     {
+        $disciplinas = DB::table('disciplinas')->where('id_professor', $id)->get();
+
+        if ($disciplinas) {
+            return redirect()->back()->withInput()->withErrors(['O usuário está associado à uma disciplina.']);
+        }
+
+        $avaliacoes = DB::table('avaliacoes')->where('id_professor', $id)->get();
+
+        if ($avaliacoes) {
+            return redirect()->back()->withInput()->withErrors(['O usuário está associado à uma avaliação.']);
+        }
+
         $user = DB::table('users')->where('id', $id)->delete();
 
         if ($user) {
